IP-425: Fixes Redis CPU counters

+++ spec/models/database_spec.rb
@@ -58,7 +58,7 @@ describe Routemaster::Models::Database do
   end
 
   describe '#used_cpu_*' do
-    it { expect(subject.used_cpu_sys).to  be_a_kind_of(Float) }
-    it { expect(subject.used_cpu_user).to be_a_kind_of(Float) }
+    it { expect(subject.used_cpu_sys).to  be_a_kind_of(Integer) }
+    it { expect(subject.used_cpu_user).to be_a_kind_of(Integer) }
   end
 end
